requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
redis[asyncio]
python-dotenv
